improved

+++ gameserver/src/main/java/com/lhf/game/creature/intelligence/choosers/TargetChooser.java
@@ -5,5 +5,5 @@ import java.util.SortedMap;
 import com.lhf.game.creature.intelligence.BasicAI.BattleMemories;
 
 public interface TargetChooser {
-    public SortedMap<String, Integer> chooseTarget(BattleMemories battleMemories);
+    public SortedMap<String, Float> chooseTarget(BattleMemories battleMemories);
 }
